Refactor: don't async load hooks
There's an argument to be made for also having an async hook provider, but the hooks themselves are async, so this is a lot of duplication for an edge case. If plugins need to do work, they can cache it themselves.

Async provision of the core MarkdownIt plugin makes sense, as these need to be registered immediately.